Update README.md
Hi, I am a student in a technical writing class. I am editing this for an assignment for class (CMPT376W) so if you find the revision useful feel free to merge it and if not let me know so I can propose more useful changes for other repos. The edit I made wasfor line 38:
1.) Starting the sentence with "Instead" but there was no alternate solution made it seem out of place so I changed it to this