Update README.md
Hi, I am a student in a technical writing class. I am editing this for an assignment for class (CMPT376W) so if you find the revision useful feel free to merge it and if not let me know so I can propose more useful changes for other repos. The edit I made wasfor line 38:
1.) Starting the sentence with "Instead" but there was no alternate solution made it seem out of place so I changed it to this
2.) On line 40 it mentions "However, we understand that this tool wouldn't be useful if you couldn't customize it when you are ready for it." Is this acknowledging your tool's limitations that it is not customizable? Or are you saying that your tool solves the problem of customization. If its the former then you can remove the "However" and say This tool is not customizable but if its the latter then I think you should provide some resources or commands that will allow someone to customize the website.